JG-24 Feat: 사이드바 UI 개발 및 스타일링 수정
