+++ src/components/SideBar/index.jsx
@@ -1,0 +1,6 @@
+import * as S from './styles';
+function SideBar() {
+  return <S.Wrapper>sdf</S.Wrapper>;
+}
+
+export default SideBar;
