+++ src/components/SideBar/styles.js
@@ -1,7 +1,8 @@
 import styled from 'styled-components';
 
 export const Wrapper = styled.div`
-  width: 100vw;
-  height: 100vh;
+  width: 390px;
+  height: 100%;
   background-color: aliceblue;
+  z-index: 2;
 `;
