+++ src/pages/mainpage/index.jsx
@@ -1,14 +1,18 @@
-import NavigationBar from '@components/NavigationBar/NavigationBar';
+import NavigationBar from '@components/NavigationBar';
+import SideBar from '@components/SideBar';
 import { NavermapsProvider } from 'react-naver-maps';
 
 import NaverMapContainer from './NaverMapContainer';
-
+import * as S from './styles';
 function Mainpage() {
   const naverMapClientId = import.meta.env.VITE_NAVER_MAP_CLIENT_ID;
 
   return (
     <>
-      <NavigationBar />
+      <S.SideWrapper>
+        <NavigationBar />
+        <SideBar />
+      </S.SideWrapper>
       <NavermapsProvider ncpClientId={naverMapClientId}>
         <NaverMapContainer />
       </NavermapsProvider>
