+++ src/pages/mainpage/styles.js
@@ -1,0 +1,13 @@
+import styled from 'styled-components';
+
+export const SideWrapper = styled.div`
+  width: 450px;
+  height: 100%;
+  box-shadow: 0 0 10px 0 rgba(0, 0, 0, 0.1);
+  position: absolute;
+  top: 0;
+  left: 0;
+  display: flex;
+  flex-direction: row;
+  z-index: 2;
+`;
